move runtime config into local_model
