+++ lib/llm/src/local_model/runtime_config.rs
